Modern Logic updates from Linea Sketch
• Write the base layer count to PSD as a positive number
• Remove method version after change to property
• Fixed version of building flags via bit shifting (comment)
• Comments to organize writing code
• Write layer data with RLE compression
• Added some comments to mirror PSD file format docs
• Remove method version of property
• Use FMUnused in another place
• Fixed analyzer warning.
• Changed to a property as it is slightly more swift-friendly
• Fixed a typo that bugged me. Added instancetype so the initializer works better with Swift.
• Fixed new iOS 11 warning deprecations
• When using dispatch_apply, make a copy of the width into a local for the block to use.
• Added ability to set the color space to FMPSD.
• Silenced remaining warnings.
• Added FMUnused and use it to tag several places where variables are used only in debug code for asserts (to silence the warnings when building release).
• Commented out NSLog where FMPSDStream would write out the number of bytes whenever it was closed.
• Close memory streams after reading the outputData from them before they are released. Not sure this is necessary, but there was a crash releasing one recorded by TestFlight, so it seems a good precaution to close them explicitly before they exit scope.
• Use `instancetype` as the return type for factory class methods in FMPSDLayer.

+++ fmpsd/classes/FMPSDUtils.m
@@ -286,10 +286,10 @@ void FMPSDDecodeRLE(char *src, int sindex, int slen, char *dst, int dindex) {
 NSString * FMPSDStringForHFSTypeCode(OSType hfsFileTypeCode) {
     
     return [NSString stringWithFormat:@"%c%c%c%c",
-                (hfsFileTypeCode >> 24) & 0xFF,
-                (hfsFileTypeCode >> 16) & 0xFF,
-                (hfsFileTypeCode >>  8) & 0xFF,
-                (hfsFileTypeCode      ) & 0xFF];
+                (int) (hfsFileTypeCode >> 24) & 0xFF,
+                (int) (hfsFileTypeCode >> 16) & 0xFF,
+                (int) (hfsFileTypeCode >>  8) & 0xFF,
+                (int) (hfsFileTypeCode      ) & 0xFF];
     
 }
